requirements.txt
torch
pillow>=9.4.0
tqdm>=4.65.0
numpy>=1.24.3
torchvision>=0.15.2
scikit-learn>=1.2.2
matplotlib>=3.7.1
seaborn>=0.11.2
torchmetrics
